refactor: Header.hook.tsx로 로직 분리
